created AddTransactionRow.tsx
