+++ components/TransactionTable.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useMemo } from "react";
 import { useState, useRef } from "react";
 import {
-  Check,
   ChevronDown,
   ChevronRight,
   FileText,
@@ -16,6 +15,7 @@ import BulkActions from "./BulkActions";
 import ContextMenu from "./ContextMenu";
 import TransactionTableHeader from "./TransactionTableHeader";
 import DriveFile, { AttachButton } from "./DriveFileCell";
+import AddTransactionRow from "./AddTransactionRow";
 
 interface TransactionTableProps {
   transactions: Transaction[];
@@ -67,11 +67,6 @@ interface TransactionTableProps {
   onToggleAbsSort: () => void;
   absValue: boolean;
 }
-
-const localToday = () => {
-  const d = new Date();
-  return `${d.getFullYear()}-${String(d.getMonth() + 1).padStart(2, "0")}-${String(d.getDate()).padStart(2, "0")}`;
-};
 
 function buildReceiptName(
   tx: { date: string; category: string | null; description: string },
@@ -138,20 +133,6 @@ const TransactionTable = ({
   onToggleAbsSort,
   absValue,
 }: TransactionTableProps) => {
-  const emptyNewTransaction: Partial<Transaction> = {
-    date: localToday(),
-    description: "",
-    category: null,
-    amount: 0,
-    status: "Completed",
-    source: null,
-    isGroup: false,
-    parentId: null,
-    driveFileId: null,
-  };
-  const [newTransaction, setNewTransaction] =
-    useState<Partial<Transaction>>(emptyNewTransaction);
-  const [showAddErrors, setShowAddErrors] = useState(false);
   const [contextMenuPos, setContextMenuPos] = useState<{
     x: number;
     y: number;
@@ -164,16 +145,6 @@ const TransactionTable = ({
     }
     setContextMenuPos({ x: e.clientX, y: e.clientY });
   };
-
-  useEffect(() => {
-    if (!showAddRow) {
-      setNewTransaction(emptyNewTransaction);
-      setShowAddErrors(false);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [showAddRow]);
-
-  const newDescriptionRef = useRef<HTMLInputElement>(null);
   const [groupDescSearch, setGroupDescSearch] = useState<
     Record<string, string>
   >({});
@@ -361,41 +332,6 @@ const TransactionTable = ({
     }
   };
 
-  const handleNewRowKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter") {
-      e.preventDefault();
-      handleSaveNew();
-    } else if (e.key === "Escape") {
-      onCancelAdd();
-    }
-  };
-
-  const handleSaveNew = async () => {
-    if (
-      !newTransaction.date ||
-      !newTransaction.description ||
-      newTransaction.amount === undefined ||
-      newTransaction.category === undefined
-    ) {
-      setShowAddErrors(true);
-      return;
-    }
-    await onAdd({
-      date: newTransaction.date,
-      description: newTransaction.description,
-      category: newTransaction.category,
-      amount: newTransaction.amount,
-      status: newTransaction.status as Status,
-      source: newTransaction.source ?? null,
-      isGroup: false,
-      parentId: null,
-      driveFileId: null,
-    });
-    setShowAddErrors(false);
-    setNewTransaction(emptyNewTransaction);
-    newDescriptionRef.current?.focus();
-  };
-
   const handleCreateGroup = async () => {
     if (!canGroup) return;
     const newGroupId = await onCreateGroup("New Group");
@@ -432,7 +368,6 @@ const TransactionTable = ({
     editingCell?.id === id && editingCell?.field === field;
 
   const tdClass = `h-9 px-4 text-[13px] border-b border-r border-gray-100 dark:border-gray-800 text-gray-600 dark:text-gray-400 whitespace-nowrap`;
-  const addInputClass = `w-full bg-transparent border-0 border-b border-transparent hover:border-gray-200 dark:hover:border-gray-700 focus:ring-0 p-1 text-[13px] text-gray-900 dark:text-foreground placeholder-gray-400 dark:placeholder-gray-500 transition-colors outline-none`;
   const editInputClass = `w-full bg-transparent border-0 outline-none text-[13px] text-gray-900 dark:text-foreground p-0 m-0 focus:ring-0 caret-gray-400 dark:caret-gray-500`;
 
   // Recursive child row renderer — handles groups at any nesting depth
@@ -839,124 +774,11 @@ const TransactionTable = ({
               </tr>
             )}
             {/* Add Transaction Row */}
-            {showAddRow && (
-              <tr
-                className="border-b border-gray-100 dark:border-gray-800"
-                onKeyDown={handleNewRowKeyDown}
-              >
-                <td className="h-9 px-3 border-r border-gray-100 dark:border-gray-800" />
-                {/* Date */}
-                <td className="h-9 px-3 border-r border-gray-100 dark:border-gray-800">
-                  <input
-                    type="date"
-                    value={newTransaction.date}
-                    className={`${addInputClass} ${showAddErrors && !newTransaction.date ? "border-rose-400! dark:border-rose-500!" : ""}`}
-                    onChange={(e) =>
-                      setNewTransaction({
-                        ...newTransaction,
-                        date: e.target.value,
-                      })
-                    }
-                    autoFocus
-                  />
-                </td>
-                {/* Description */}
-                <td className="h-9 px-3 border-r border-gray-100 dark:border-gray-800">
-                  <input
-                    ref={newDescriptionRef}
-                    type="text"
-                    placeholder="Description..."
-                    value={newTransaction.description}
-                    className={`${addInputClass} ${showAddErrors && !newTransaction.description ? "border-rose-400! dark:border-rose-500!" : ""}`}
-                    onChange={(e) =>
-                      setNewTransaction({
-                        ...newTransaction,
-                        description: e.target.value,
-                      })
-                    }
-                  />
-                </td>
-                {/* Category */}
-                <td className="h-9 px-3 border-r border-gray-100 dark:border-gray-800">
-                  <input
-                    type="text"
-                    placeholder="Category..."
-                    value={newTransaction.category ?? ""}
-                    autoCapitalize="none"
-                    className={addInputClass}
-                    onChange={(e) =>
-                      setNewTransaction({
-                        ...newTransaction,
-                        category: e.target.value || null,
-                      })
-                    }
-                  />
-                </td>
-                {/* Amount */}
-                <td className="h-9 px-3 border-r border-gray-100 dark:border-gray-800">
-                  <input
-                    type="number"
-                    placeholder="0.00"
-                    step="0.01"
-                    value={newTransaction.amount || ""}
-                    className={`${addInputClass} text-right ${showAddErrors && newTransaction.amount === undefined ? "border-rose-400! dark:border-rose-500!" : ""}`}
-                    onChange={(e) =>
-                      setNewTransaction({
-                        ...newTransaction,
-                        amount: parseFloat(e.target.value),
-                      })
-                    }
-                  />
-                </td>
-                {/* Status */}
-                <td className="h-9 px-3 border-r border-gray-100 dark:border-gray-800">
-                  <select
-                    value={newTransaction.status}
-                    className={addInputClass}
-                    onChange={(e) =>
-                      setNewTransaction({
-                        ...newTransaction,
-                        status: e.target.value as Status,
-                      })
-                    }
-                  >
-                    {STATUSES.map((s) => (
-                      <option key={s} value={s}>
-                        {s}
-                      </option>
-                    ))}
-                  </select>
-                </td>
-                {/* Source */}
-                <td className="h-9 px-3 border-r border-gray-100 dark:border-gray-800">
-                  <input
-                    type="text"
-                    placeholder="Source..."
-                    value={newTransaction.source ?? ""}
-                    autoCapitalize="none"
-                    className={addInputClass}
-                    onChange={(e) =>
-                      setNewTransaction({
-                        ...newTransaction,
-                        source: e.target.value || null,
-                      })
-                    }
-                  />
-                </td>
-                {/* File — not available on new row */}
-                <td className="h-9 px-3 border-r border-gray-100 dark:border-gray-800" />
-                {/* Extra */}
-                <td className={tdClass}>
-                  <button
-                    onClick={handleSaveNew}
-                    aria-label="Save Transaction"
-                    className="p-1 text-gray-400 hover:text-emerald-600 dark:text-gray-500 dark:hover:text-emerald-400 transition-colors"
-                  >
-                    <Check className="w-4 h-4" />
-                  </button>
-                </td>
-              </tr>
-            )}
+            <AddTransactionRow
+              showAddRow={showAddRow}
+              onAdd={onAdd}
+              onCancelAdd={onCancelAdd}
+            />
 
             {/* Transaction rows */}
             {transactions.length === 0 ? (
